# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creators Dashboard evaluator API
  description: |
    A Flask + MongoDB API for the evaluator service.
    
    ## Authentication
    All endpoints (except `/dev-login` and `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a token from the `/dev-login` endpoint.
  version: 0.1.0
  contact:
    name: Agile Crafts People
    email: info@agilecrafts.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:9098
    description: Local development server
...
paths:
...
  /api/Grade:
    get:
      tags:
        - Grade
      summary: Get all Grade documents
      description: Retrieve infinite scroll batch of sorted, filtered Grade documents
      operationId: getGrades
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Grade
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, created.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved Grades
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Grade'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Grade
      summary: Create a new Grade document
      description: Create a new Grade document (requires staff or admin role)
      operationId: createGrade
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeInput'
      responses:
        '201':
          description: Grade created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Grade/{GradeId}:
    get:
      tags:
        - Grade
      summary: Get Grade by GradeId
      description: Retrieve a specific Grade document by its GradeId
      operationId: getGrade
      security:
        - bearerAuth: []
      parameters:
        - name: GradeId
          in: path
          required: true
          description: The Grade ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Grade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/dev-login` endpoint.
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123
...
    Grade:
      type: object
      description: Grade domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Grade name
          pattern: '^[^\s]{1,40}$'
          example: my-Grade
        description:
          type: string
          description: Grade description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Grade document
        status:
          type: string
          description: Grade status
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
    GradeInput:
      type: object
      description: Input for creating a Grade document
      properties:
        name:
          type: string
          description: Grade name
          pattern: '^[^\s]{1,40}$'
          example: my-Grade
        description:
          type: string
          description: Grade description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Grade document
        status:
          type: string
          description: Grade status
          example: active
...